pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-predictive-intelligence"
version = "0.2.0"
description = "Multi-modal AI platform for market prediction using time-series, NLP, and fusion models"
requires-python = ">=3.10"
dependencies = [
    "pandas",
    "numpy",
    "scipy",
    "requests",
    "praw",
    "yfinance",
    "fredapi",
    "pycoingecko",
    "beautifulsoup4",
    "kaggle",
    "python-dotenv",
    "tqdm",
    "pytrends",
    "nltk",
    "torch",
    "transformers",
    "scikit-learn",
    "joblib",
    "pyyaml",
    "schedule",
    "google-search-results",
    "pyarrow",
    "duckdb",
    "streamlit",
    "plotly",
    "optuna",
    "mlflow",
    "shap",
    "aiohttp",
    "matplotlib",
    "apscheduler",
    "xmltodict",
    "hmmlearn",
]

[project.optional-dependencies]
test = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
